Clean up Conversations logo SVG
The previous version of the logo had some wonky alignments and other
artifacts in some places on the dashed line. This commit provides a new
version of the logo with perfect alignments.

It also includes a single path for the dashed line, which can be used
to ease further modifications to the logo.