Add referrer to built-in snikket.org website links
This helps identify visitors coming to the website from the
app. I suspect a lot of people install the app, tap 'Learn more'
and end their journey after being overwhelmed by that page.

I'd like to confirm that theory, and possibly make a new
dedicated page for the in-app link to point to.