downgrade webrtc to m104
m107 shipped with 2.11.0 (both fdroid and play) was causing problems when
calling between some (not all) devices.

The callee (repsonder) would not see the video of the caller (initiator).

Our best guess is that this has something to do with the new av1 decoder and it
only occurs between devices where av1 is selected as the codec.

It's probably selected on 'modern' devices. It's not happening between a
Pixel 4a and a Xiaomi Mi A1 but it is happening between two Pixels