Remove READ_CONTACTS permission from Play Store version
Google has been flagging all XMPP apps with this permission, claiming that
they upload users' contacts to a third-party without properly disclosing such.

The reality is that the contacts integration is actually entirely local, and
allows you to (optionally) link your XMPP contacts with entries in your phone
contacts, so you can see your preferred contact photo, etc.

As nice a little feature as this is, it's probably not widely used, and it's
not worth having Google delist the app entirely due to it.