play ringtone through notification
in 484f63318081199a3d80325ae984f50ef66acd0b we switched from letting the notification play the ringtone and handle the vibration to doing it ourselves.

this had two reasons. The ringtone selector in the android notification settings was broken at times and we wanted to silence the notification when pressing volume down.

This commit essentially reverts that change. We fixed the ringtone selection by handling it internally in Conversations and simply recreating the entire channel. Silencing the notification can be achieved by re-posting it with onlyAlertOnce set to true (I guess we didn’t know that at the time)