ensure that message is parsed into bubble OR receipt, marker, reaction
This commit ensures that Conversations always parses a message into a
bubble or as meta data (receipt, display marker, reaction) never both
and with a preference for bubble.

Conversations goes through great length to ensure that all participants
of a group chat have a similiar (or ideally the same) view of the chat
history.
For example Conversations doesn’t use xhtml, shows a language tag when
the content is available in multiple languages, etc.

So a message that has both a body and a reaction for example shouldn't
show differently in clients that support reactions and clients that
don't.

Take the following example:

```xml
<message to='romeo@capulet.net/troll' type='groupchat'>
  <body>Conversations users are stupid!</body>
  <reactions id='a1' xmlns='urn:xmpp:reactions:0'>
    <reaction>💩</reaction>
  </reactions>
</message>
```

This message should render as what it is: A message and the body
shouldn't be silently discarded for clients that support reactions.

(The same goes for display markers and receipts)

Fallback messages for something that is not supposed to render as a
bubble should be discouraged.